requirements.txt
# Code analysis dependencies
json_strong_typing
pysqlsync

# Documentation dependencies
sphinx
sphinx-design
sphinx-rtd-theme
